Add code via upload
These are three markdowns which must be run in order, and the functions that they use to do what they do. Assuming working directory is set to the root directory, they should be runable immediately.

Note: The code uses pacman to load packages. If you do not have pacman installed, it will not run.