Added plots of results, stan files and cleaned data.
Added stan files (which are necessary to run the code). For good measure, I also added the result-plots I got when I ran the code, as well as the cleaned data (so users need not run the 1st notebook).

The raw numbers (draws and summaries) are too large for github, so the plots will have to do.